\name{apng-package}
\alias{apng-package}
\docType{package}
\title{
\packageTitle{apng}
}
\description{
\packageDescription{apng}
}
\author{
\packageAuthor{apng}

Maintainer: \packageMaintainer{apng}
}
\references{\url{https://www.w3.org/TR/PNG/}\cr
\url{https://wiki.mozilla.org/APNG_Specification}
}
\note{The CRC implementation in this package was adopted from the W3 Portable Network Graphics (PNG) Specification (Second Edition): Annex D - Sample Cyclic Redundancy Code implementation.
As such, special thanks go out to the authors of the specification: \url{https://www.w3.org/TR/PNG/#F-Relationship}.}
